move CheckpointStore to log.py
